requirements.txt
arrow==1.3.0
asgiref==3.8.1
babel==2.17.0
binaryornot==0.4.4
black==25.1.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cookiecutter==2.6.0
crispy-bootstrap4==2024.10
dj-database-url==3.0.0
Django==5.2.1
django-adminlte3==0.1.6
django-appconf==1.1.0
django-browser-reload==1.18.0
django-compressor==4.5.1
django-crispy-forms==2.4
django-filter==25.1
django-tailwind==4.0.1
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
flake8==7.2.0
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.8
psycopg2==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
rcssmin==1.1.2
requests==2.32.3
rich==14.0.0
rjsmin==1.2.2
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
XlsxWriter==3.2.3
